fix type checking

+++ toxcore/tox.h
@@ -445,7 +445,7 @@ void tox_callback_group_action(Tox *tox, void (*function)(Tox *tox, int, int, co
  * Function(Tox *tox, int groupnumber, int peernumber, uint8_t * title, uint8_t length, void *userdata)
  * if peernumber == -1, then author is unknown (e.g. initial joining the group)
  */
-void tox_callback_group_title(Tox *tox, void (*function)(Tox *tox, int, int, const uint8_t *, uint16_t,
+void tox_callback_group_title(Tox *tox, void (*function)(Tox *tox, int, int, const uint8_t *, uint8_t,
                              void *), void *userdata);
 
 /* Set callback function for peer name list changes.
